Refactor effect for fetching stats into a hook
