+++ webapp/src/components/backstage/playbooks/playbook.tsx
@@ -16,15 +16,14 @@ import {GlobalState} from 'mattermost-redux/types/store';
 import {useIntl} from 'react-intl';
 
 import {navigateToUrl, navigateToPluginUrl, pluginErrorUrl} from 'src/browser_routing';
-import {useExperimentalFeaturesEnabled, useForceDocumentTitle} from 'src/hooks';
+import {useExperimentalFeaturesEnabled, useForceDocumentTitle, useStats} from 'src/hooks';
 import PlaybookUsage from 'src/components/backstage/playbooks/playbook_usage';
 import PlaybookPreview from 'src/components/backstage/playbooks/playbook_preview';
 
 import {SecondaryButtonLargerRight} from 'src/components/backstage/playbook_runs/shared';
-import {clientFetchPlaybook, telemetryEventForPlaybook, fetchPlaybookStats} from 'src/client';
+import {clientFetchPlaybook, telemetryEventForPlaybook} from 'src/client';
 import {ErrorPageTypes} from 'src/constants';
 import {PlaybookWithChecklist} from 'src/types/playbook';
-import {EmptyPlaybookStats} from 'src/types/stats';
 import {PrimaryButton} from 'src/components/assets/buttons';
 import ClipboardsPlay from 'src/components/assets/icons/clipboards_play';
 import {RegularHeading} from 'src/styles/headings';
@@ -46,7 +45,7 @@ const Playbook = () => {
     const [playbook, setPlaybook] = useState<PlaybookWithChecklist | null>(null);
     const [fetchingState, setFetchingState] = useState(FetchingStateType.loading);
     const team = useSelector<GlobalState, Team>((state) => getTeam(state, playbook?.team_id || ''));
-    const [stats, setStats] = useState(EmptyPlaybookStats);
+    const stats = useStats(match.params.playbookId);
 
     useForceDocumentTitle(playbook?.title ? (playbook.title + ' - Playbooks') : 'Playbooks');
 
@@ -85,20 +84,6 @@ const Playbook = () => {
         };
 
         fetchData();
-    }, [match.params.playbookId]);
-
-    useEffect(() => {
-        const fetchStats = async () => {
-            try {
-                const ret = await fetchPlaybookStats(match.params.playbookId);
-                setStats(ret);
-            } catch {
-                // Ignore any errors here. If it fails, it's most likely also failed to fetch
-                // the playbook above.
-            }
-        };
-
-        fetchStats();
     }, [match.params.playbookId]);
 
     if (fetchingState === FetchingStateType.loading) {
